Refactoring: Move database strcuts to db package
This allows usage of db from diffrent services withouth causing cycle
dependencies
