+++ identityservice/contract/organizationsglobalidcontracts_api.go
@@ -3,6 +3,7 @@ package contract
 import (
 	"encoding/json"
 	"net/http"
+	"github.com/itsyouonline/identityserver/db/contract"
 )
 
 type OrganizationsglobalidcontractsAPI struct {
@@ -12,7 +13,7 @@ type OrganizationsglobalidcontractsAPI struct {
 // date.
 // It is handler for GET /organizations/{globalid}/contracts
 func (api OrganizationsglobalidcontractsAPI) Get(w http.ResponseWriter, r *http.Request) { // includeExpired := req.FormValue("includeExpired")// max := req.FormValue("max")// start := req.FormValue("start")
-	var respBody Contract
+	var respBody contract.Contract
 	json.NewEncoder(w).Encode(&respBody)
 	// uncomment below line to add header
 	// w.Header().Set("key","value")
