+++ identityservice/user/users_api.go
@@ -10,6 +10,7 @@ import (
 	"github.com/itsyouonline/identityserver/identityservice/contract"
 	"github.com/itsyouonline/identityserver/identityservice/invitations"
 	"github.com/itsyouonline/identityserver/credentials/password"
+	"github.com/itsyouonline/identityserver/db/user"
 	"github.com/itsyouonline/identityserver/db/user/apikey"
 )
 
@@ -19,14 +20,14 @@ type UsersAPI struct {
 // It is handler for POST /users
 func (api UsersAPI) Post(w http.ResponseWriter, r *http.Request) {
 
-	var u User
+	var u user.User
 
 	if err := json.NewDecoder(r.Body).Decode(&u); err != nil {
 		http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
 		return
 	}
 
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 	if err := userMgr.Save(&u); err != nil {
 		log.Error("ERROR while saving user:\n", err)
 		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
@@ -43,7 +44,7 @@ func (api UsersAPI) Post(w http.ResponseWriter, r *http.Request) {
 func (api UsersAPI) usernameGet(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
 
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	user, err := userMgr.GetByName(username)
 	if err != nil {
@@ -85,7 +86,7 @@ func (api UsersAPI) RegisterNewEmailAddress(w http.ResponseWriter, r *http.Reque
 		return
 	}
 
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 	u, err := userMgr.GetByName(username)
 	if err != nil {
 		log.Error(err)
@@ -131,7 +132,7 @@ func (api UsersAPI) UpdateEmailAddress(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	if oldlabel != body.Label {
 		u, err := userMgr.GetByName(username)
@@ -172,7 +173,7 @@ func (api UsersAPI) DeleteEmailAddress(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
 	label := mux.Vars(r)["label"]
 
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	u, err := userMgr.GetByName(username)
 	if err != nil {
@@ -203,7 +204,7 @@ func (api UsersAPI) DeleteEmailAddress(w http.ResponseWriter, r *http.Request) {
 // Delete the associated Github account.
 func (api UsersAPI) DeleteGithubAccount(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 	err := userMgr.DeleteGithubAccount(username)
 	if err != nil {
 		log.Error(err)
@@ -219,7 +220,7 @@ func (api UsersAPI) DeleteGithubAccount(w http.ResponseWriter, r *http.Request) 
 func (api UsersAPI) DeleteFacebookAccount(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
 
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 	err := userMgr.DeleteFacebookAccount(username)
 	if err != nil {
 		log.Error(err)
@@ -242,7 +243,7 @@ func (api UsersAPI) UpdatePassword(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
 		return
 	}
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 	exists, err := userMgr.Exists(username)
 	if ! exists || err != nil  {
 		http.Error(w, http.StatusText(http.StatusNotFound), http.StatusNotFound)
@@ -267,9 +268,9 @@ func (api UsersAPI) UpdatePassword(w http.ResponseWriter, r *http.Request) {
 // GetUserInformation is the handler for GET /users/{username}/info
 func (api UsersAPI) GetUserInformation(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
-	user, err := userMgr.GetByName(username)
+	userobj, err := userMgr.GetByName(username)
 	if err != nil {
 		http.Error(w, http.StatusText(http.StatusNotFound), http.StatusNotFound)
 		return
@@ -288,21 +289,21 @@ func (api UsersAPI) GetUserInformation(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	respBody := &userview{
-		Username: user.Username,
+	respBody := &Userview{
+		Username: userobj.Username,
 	}
 
 	if authorization.Github {
-		respBody.Github = user.Github.Name
+		respBody.Github = userobj.Github.Name
 	}
 	if authorization.Facebook {
-		respBody.Facebook = user.Facebook.Name
+		respBody.Facebook = userobj.Facebook.Name
 	}
 	if authorization.Address != nil {
-		respBody.Address = make(map[string]Address)
+		respBody.Address = make(map[string]user.Address)
 
 		for requestedLabel, realLabel := range authorization.Address {
-			if value, found := user.Address[realLabel]; found {
+			if value, found := userobj.Address[realLabel]; found {
 				respBody.Address[requestedLabel] = value
 			}
 		}
@@ -312,27 +313,27 @@ func (api UsersAPI) GetUserInformation(w http.ResponseWriter, r *http.Request) {
 		respBody.Email = make(map[string]string)
 
 		for requestedLabel, realLabel := range authorization.Email {
-			if value, found := user.Email[realLabel]; found {
+			if value, found := userobj.Email[realLabel]; found {
 				respBody.Email[requestedLabel] = value
 			}
 		}
 	}
 
 	if authorization.Phone != nil {
-		respBody.Phone = make(map[string]Phonenumber)
+		respBody.Phone = make(map[string]user.Phonenumber)
 
 		for requestedLabel, realLabel := range authorization.Phone {
-			if value, found := user.Phone[realLabel]; found {
+			if value, found := userobj.Phone[realLabel]; found {
 				respBody.Phone[requestedLabel] = value
 			}
 		}
 	}
 
 	if authorization.Bank != nil {
-		respBody.Bank = make(map[string]BankAccount)
+		respBody.Bank = make(map[string]user.BankAccount)
 
 		for requestedLabel, realLabel := range authorization.Bank {
-			if value, found := user.Bank[realLabel]; found {
+			if value, found := userobj.Bank[realLabel]; found {
 				respBody.Bank[requestedLabel] = value
 			}
 		}
@@ -355,7 +356,7 @@ func (api UsersAPI) usernamevalidateGet(w http.ResponseWriter, r *http.Request) 
 // Register a new phonenumber
 func (api UsersAPI) RegisterNewPhonenumber(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	u, err := userMgr.GetByName(username)
 	if err != nil {
@@ -365,7 +366,7 @@ func (api UsersAPI) RegisterNewPhonenumber(w http.ResponseWriter, r *http.Reques
 
 	body := struct {
 		Label       string
-		Phonenumber Phonenumber
+		Phonenumber user.Phonenumber
 	}{}
 
 	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
@@ -406,7 +407,7 @@ func (api UsersAPI) RegisterNewPhonenumber(w http.ResponseWriter, r *http.Reques
 // usernamephonenumbersGet is the handler for GET /users/{username}/phonenumbers
 func (api UsersAPI) usernamephonenumbersGet(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	user, err := userMgr.GetByName(username)
 	if err != nil {
@@ -422,21 +423,21 @@ func (api UsersAPI) usernamephonenumbersGet(w http.ResponseWriter, r *http.Reque
 func (api UsersAPI) usernamephonenumberslabelGet(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
 	label := mux.Vars(r)["label"]
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
-	user, err := userMgr.GetByName(username)
+	userobj, err := userMgr.GetByName(username)
 	if err != nil {
 		http.Error(w, http.StatusText(http.StatusNotFound), http.StatusNotFound)
 		return
 	}
 
-	if _, ok := user.Phone[label]; ok != true {
+	if _, ok := userobj.Phone[label]; ok != true {
 		http.Error(w, http.StatusText(http.StatusNotFound), http.StatusNotFound)
 		return
 	}
 
-	respBody := map[string]Phonenumber{
-		label: user.Phone[label],
+	respBody := map[string]user.Phonenumber{
+		label: userobj.Phone[label],
 	}
 
 	w.Header().Set("Content-Type", "application/json")
@@ -451,7 +452,7 @@ func (api UsersAPI) UpdatePhonenumber(w http.ResponseWriter, r *http.Request) {
 
 	body := struct {
 		Label       string
-		Phonenumber Phonenumber
+		Phonenumber user.Phonenumber
 	}{}
 
 	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
@@ -469,7 +470,7 @@ func (api UsersAPI) UpdatePhonenumber(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	u, err := userMgr.GetByName(username)
 	if err != nil {
@@ -514,7 +515,7 @@ func (api UsersAPI) UpdatePhonenumber(w http.ResponseWriter, r *http.Request) {
 func (api UsersAPI) DeletePhonenumber(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
 	label := mux.Vars(r)["label"]
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	user, err := userMgr.GetByName(username)
 	if err != nil {
@@ -540,11 +541,11 @@ func (api UsersAPI) DeletePhonenumber(w http.ResponseWriter, r *http.Request) {
 // It is handler for POST /users/{username}/banks
 func (api UsersAPI) usernamebanksPost(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	body := struct {
 		Label string
-		Bank  BankAccount
+		Bank  user.BankAccount
 	}{}
 
 	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
@@ -586,7 +587,7 @@ func (api UsersAPI) usernamebanksPost(w http.ResponseWriter, r *http.Request) {
 // It is handler for GET /users/{username}/banks
 func (api UsersAPI) usernamebanksGet(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	user, err := userMgr.GetByName(username)
 	if err != nil {
@@ -602,21 +603,21 @@ func (api UsersAPI) usernamebanksGet(w http.ResponseWriter, r *http.Request) {
 func (api UsersAPI) usernamebankslabelGet(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
 	label := mux.Vars(r)["label"]
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
-	user, err := userMgr.GetByName(username)
+	userobj, err := userMgr.GetByName(username)
 	if err != nil {
 		http.Error(w, http.StatusText(http.StatusNotFound), http.StatusNotFound)
 		return
 	}
 
-	if _, ok := user.Bank[label]; ok != true {
+	if _, ok := userobj.Bank[label]; ok != true {
 		http.Error(w, http.StatusText(http.StatusNotFound), http.StatusNotFound)
 		return
 	}
 
-	respBody := map[string]BankAccount{
-		label: user.Bank[label],
+	respBody := map[string]user.BankAccount{
+		label: userobj.Bank[label],
 	}
 
 	w.Header().Set("Content-Type", "application/json")
@@ -628,11 +629,11 @@ func (api UsersAPI) usernamebankslabelGet(w http.ResponseWriter, r *http.Request
 func (api UsersAPI) usernamebankslabelPut(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
 	oldlabel := mux.Vars(r)["label"]
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	body := struct {
 		Label string
-		Bank  BankAccount
+		Bank  user.BankAccount
 	}{}
 
 	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
@@ -686,7 +687,7 @@ func (api UsersAPI) usernamebankslabelPut(w http.ResponseWriter, r *http.Request
 func (api UsersAPI) usernamebankslabelDelete(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
 	label := mux.Vars(r)["label"]
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	user, err := userMgr.GetByName(username)
 	if err != nil {
@@ -712,11 +713,11 @@ func (api UsersAPI) usernamebankslabelDelete(w http.ResponseWriter, r *http.Requ
 // Register a new address
 func (api UsersAPI) RegisterNewAddress(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	body := struct {
 		Label   string
-		Address Address
+		Address user.Address
 	}{}
 
 	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
@@ -758,7 +759,7 @@ func (api UsersAPI) RegisterNewAddress(w http.ResponseWriter, r *http.Request) {
 // It is handler for GET /users/{username}/addresses
 func (api UsersAPI) usernameaddressesGet(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	user, err := userMgr.GetByName(username)
 	if err != nil {
@@ -774,21 +775,21 @@ func (api UsersAPI) usernameaddressesGet(w http.ResponseWriter, r *http.Request)
 func (api UsersAPI) usernameaddresseslabelGet(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
 	label := mux.Vars(r)["label"]
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
-	user, err := userMgr.GetByName(username)
+	userobj, err := userMgr.GetByName(username)
 	if err != nil {
 		http.Error(w, http.StatusText(http.StatusNotFound), http.StatusNotFound)
 		return
 	}
 
-	if _, ok := user.Address[label]; ok != true {
+	if _, ok := userobj.Address[label]; ok != true {
 		http.Error(w, http.StatusText(http.StatusNotFound), http.StatusNotFound)
 		return
 	}
 
-	respBody := map[string]Address{
-		label: user.Address[label],
+	respBody := map[string]user.Address{
+		label: userobj.Address[label],
 	}
 
 	w.Header().Set("Content-Type", "application/json")
@@ -803,7 +804,7 @@ func (api UsersAPI) UpdateAddress(w http.ResponseWriter, r *http.Request) {
 
 	body := struct {
 		Label   string
-		Address Address
+		Address user.Address
 	}{}
 
 	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
@@ -816,7 +817,7 @@ func (api UsersAPI) UpdateAddress(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	u, err := userMgr.GetByName(username)
 	if err != nil {
@@ -860,7 +861,7 @@ func (api UsersAPI) UpdateAddress(w http.ResponseWriter, r *http.Request) {
 func (api UsersAPI) DeleteAddress(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
 	label := mux.Vars(r)["label"]
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	u, err := userMgr.GetByName(username)
 	if err != nil {
@@ -929,7 +930,7 @@ func (api UsersAPI) usernameorganizationsGet(w http.ResponseWriter, r *http.Requ
 func (api UsersAPI) GetAllAuthorizations(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
 
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	authorizations, err := userMgr.GetAuthorizationsByUser(username)
 	if err != nil {
@@ -948,7 +949,7 @@ func (api UsersAPI) GetAuthorization(w http.ResponseWriter, r *http.Request) {
 	username := mux.Vars(r)["username"]
 	grantedTo := mux.Vars(r)["grantedTo"]
 
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	authorization, err := userMgr.GetAuthorization(username, grantedTo)
 	if err != nil {
@@ -969,7 +970,7 @@ func (api UsersAPI) UpdateAuthorization(w http.ResponseWriter, r *http.Request) 
 	username := mux.Vars(r)["username"]
 	grantedTo := mux.Vars(r)["grantedTo"]
 
-	authorization := &Authorization{}
+	authorization := &user.Authorization{}
 
 	if err := json.NewDecoder(r.Body).Decode(authorization); err != nil {
 		http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
@@ -979,7 +980,7 @@ func (api UsersAPI) UpdateAuthorization(w http.ResponseWriter, r *http.Request) 
 	authorization.Username = username
 	authorization.GrantedTo = grantedTo
 
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	err := userMgr.UpdateAuthorization(authorization)
 	if err != nil {
@@ -998,7 +999,7 @@ func (api UsersAPI) DeleteAuthorization(w http.ResponseWriter, r *http.Request) 
 	username := mux.Vars(r)["username"]
 	grantedTo := mux.Vars(r)["grantedTo"]
 
-	userMgr := NewManager(r)
+	userMgr := user.NewManager(r)
 
 	err := userMgr.DeleteAuthorization(username, grantedTo)
 	if err != nil {
