+++ identityservice/user/userview.go
@@ -1,13 +1,15 @@
 package user
 
-type userview struct {
-	Address       map[string]Address     `json:"address"`
-	Bank          map[string]BankAccount `json:"bank"`
+import "github.com/itsyouonline/identityserver/db/user"
+
+type Userview struct {
+	Address       map[string]user.Address     `json:"address"`
+	Bank          map[string]user.BankAccount `json:"bank"`
 	Email         map[string]string      `json:"email"`
 	Facebook      string                 `json:"facebook"`
 	Github        string                 `json:"github"`
 	Organizations []string               `json:"organizations"`
-	Phone         map[string]Phonenumber `json:"phone"`
+	Phone         map[string]user.Phonenumber `json:"phone"`
 	PublicKeys    []string               `json:"publicKeys"`
 	Username      string                 `json:"username"`
 }
